[INLONG-2610][Manager] Refactor the StreamSink related class and service unit tests
